Django/requirementss.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==22.1.0
aiogram==2.15
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
asgiref==3.5.0
astroid==2.9.3
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
Babel==2.10.3
backports.zoneinfo==0.2.1
bidict==0.22.0
boto3==1.24.84
botocore==1.27.84
CacheControl==0.12.11
cachetools==4.2.4
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
cloudpickle==1.6.0
colorama==0.4.4
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==38.0.1
cycler==0.11.0
dask==2022.2.0
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.16
django-allauth==0.51.0
django-cors-headers @ file:///home/conda/feedstock_root/build_artifacts/django-cors-headers_1654546143857/work
django-filter==21.1
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
dm-tree==0.1.7
dnspython==1.16.0
docopt==0.6.2
exceptiongroup==1.0.0rc9
exotel==0.1.5
fbmessenger==6.0.0
fire==0.4.0
flatbuffers==2.0.7
fonttools==4.29.1
fsspec==2022.8.2
future==0.18.2
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.43.0
gunicorn==20.1.0
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.20.4
httptools==0.5.0
httpx==0.15.4
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.0
importlib-resources==5.9.0
isort==5.10.1
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
libclang==14.0.6
locket==1.0.0
Markdown==3.3.6
matplotlib==3.3.4
mattermostwrapper==2.2
mccabe==0.6.1
MouseInfo==0.1.3
msgpack==1.0.4
mssql-django==1.1.3
multidict==5.0.0
mysqlclient==2.0.1
networkx==2.5.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
outcome==1.2.0
packaging==20.9
pamqp==2.3.0
pandas==1.3.5
partd==1.3.0
Pillow==9.0.1
platformdirs==2.5.0
prompt-toolkit==2.0.10
protobuf==3.19.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.2
PyGetWindow==0.0.9
PyJWT==2.5.0
pykwalify==1.8.0
pylint==2.12.2
pymongo==3.10.1
PyMsgBox==1.0.9
pyodbc==4.0.34
pyparsing==3.0.7
pyperclip==1.8.2
pyreadline==2.1
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.4
python-socketio==5.7.1
python3-openid==3.2.0
pytweening==1.0.4
pytz==2021.3
PyYAML==6.0
questionary==1.9.0
randomname==0.1.5
rasa==2.8.7
rasa-sdk==2.8.6
redis==3.5.3
regex==2021.7.6
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rocketchat-API==1.16.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sanic==20.9.1
Sanic-Cors==0.10.0.post3
sanic-jwt==1.5.0
Sanic-Plugins-Framework==0.9.5
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.11.2
selenium==4.5.0
sentry-sdk==1.2.0
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
tabulate==0.8.10
tarsafe==0.0.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.4
tensorflow-addons==0.13.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-probability==0.13.0
termcolor==2.0.1
terminaltables==3.1.10
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
trio==0.22.0
trio-websocket==0.9.2
twilio==6.50.1
typed-ast==1.5.2
typeguard==2.13.3
types-cryptography==3.3.23
typing-extensions==3.10.0.2
typing-utils==0.1.0
tzdata==2022.7
tzlocal==2.1
ujson==4.3.0
urllib3==1.26.8
wcwidth==0.2.5
webdriver-manager==3.8.4
webexteamssdk==1.6.1
websockets==8.1
Werkzeug==2.0.3
whitenoise==6.0.0
wincertstore==0.2
wrapt==1.13.3
wsproto==1.2.0
yarl==1.8.1
zipp==3.7.0
